perf(torch_ops): gate V4.1 Triton backends on measured GB10 crossovers
Benchmark-driven backend selection for the DeepSeek-V4.1 ops (new harness
meta/benchmarks/bench_v41_triton.py, mirroring bench_roofline.py honesty
rules):

* dsa_indexer: reference (cuBLAS einsum) is the default backend on GPU —
  the fused Triton kernel measured ~6x slower at decode shapes on GB10
  (q[1,1,32,128], T=4096: 230 us vs 37 us); opt back in with
  VKERNELS_DSA_INDEXER_BACKEND=triton.
* fp8_gemm: add a decode-shape (M <= block) Triton block-scaled kernel with
  bit-exact e4m3 decode; default "auto" dispatches to it only at the
  measured crossover (N >= 2048: wo_b 3.15x, wq_b 1.5x; small-N loses 11x).
  Backend via VKERNELS_V41_FP8_GEMM_BACKEND.
* sparse_attention: shape-gate on the measured crossover (kernel wins at
  S <= 2, loses beyond: S=64 reference 3.2x faster) and drop BT 64->32;
  VKERNELS_V41_SPARSE_ATTN_BACKEND forces triton/reference.
* mxfp4_dequant: int64 row offsets — the flattened full-stack shape
  (884_736 x 5120) overflows int32 byte offsets (illegal memory access).

Rebased over 63a3c2d (shared _fastpath gate): backend gates compose with
fast_path() and the pow-2 shape gates instead of duplicating the probe.

Also fixes a latent defect in 63a3c2d's non-pow-2 regression test: it
asserted torch.equal between a GPU result and a CPU reference, which only
held on the author's no-GPU box (CUDA tests skipped); any CUDA machine
sees last-bit cuBLAS-vs-CPU drift. Compare against the reference on the
same device instead.

Tests cover the Triton paths (forced + auto), the padded-M masking and
per-tile B-scale bugs hit at bring-up, and the int32-overflow regression.
36/36 v41 tests pass on GB10 (bench-venv, CUDA); full suite delta vs
origin/main: +4 passed, ruff output identical.

+++ src/python/vkernels/torch_ops/v41_mxfp4_dequant.py
@@ -71,7 +71,9 @@ def _kernel():
 
     @triton.jit
     def _mxfp4(PACKED, SCALE, TABLE, Y, I: tl.constexpr, GROUP: tl.constexpr, BLOCK: tl.constexpr):
-        row = tl.program_id(0)
+        # int64: the flattened full-stack shape (E*O rows, e.g. 884_736 x 5120)
+        # overflows int32 byte offsets.
+        row = tl.program_id(0).to(tl.int64)
         col = tl.program_id(1) * BLOCK + tl.arange(0, BLOCK)  # output column in [0, I)
         mask = col < I
         byte = tl.load(PACKED + row * (I // 2) + col // 2, mask, 0).to(tl.int32)
